wordcloud moved to /frontend/
