+++ src/main/java/dk/kb/netarchivesuite/solrwayback/service/SolrWaybackResource.java
@@ -216,21 +216,6 @@ public class SolrWaybackResource {
   
   
   
-  @GET
-  @Path("/wordcloud/domain")
-  @Produces("image/png")
-  public Response  wordCloudForDomain(@QueryParam("domain") String domain) throws SolrWaybackServiceException {
-    try {                        
-        BufferedImage image = Facade.wordCloudForDomain(domain);           
-        //In jersey images could just be returned as bufferedimage class. in jax-res need to convert them manual-
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        ImageIO.write(image, "png", baos);
-        byte[] imageData = baos.toByteArray();        
-        return Response.ok(imageData).build();     
-    } catch (Exception e) {           
-      throw handleServiceExceptions(e);
-    }
-  }
   
   /*
    *
